# spring/src/main/java/itmo/ivank/soa/dto/OrganizationsPage.java
package itmo.ivank.soa.dto;

import itmo.ivank.soa.entity.Organization;

import java.util.List;

public record OrganizationsPage(
        List<Organization> organizations,
        Integer page,
        Integer size,
        Long totalElements,
        Integer totalPages
) {
}